+++ tests/math/testProfile.cpp
@@ -44,7 +44,7 @@ using namespace DGtal;
 ///////////////////////////////////////////////////////////////////////////////
 
 struct LogFct{
-  float operator()(const float &a) const {
+  double operator()(const double &a) const {
     return log(a);
   }
 };
@@ -62,8 +62,8 @@ TEST_CASE( "Testing Profile" )
       sp.addValue(2, 10.0);
       sp.addValue(3, 10.0);
       sp.addValue(4, 10.0);
-      std::vector<float> x;
-      std::vector<float> y;
+      std::vector<double> x;
+      std::vector<double> y;
       sp.getProfile(x, y);
       REQUIRE( x[3] == Approx(log(4)) );
       REQUIRE( y[3] == Approx(log(10.0)) );
@@ -81,8 +81,8 @@ TEST_CASE( "Testing Profile" )
       sp2.addValue(1, 2);
       sp2.addValue(1, 1);
       sp2.addValue(2, 4);
-      std::vector<float> x;
-      std::vector<float> y;
+      std::vector<double> x;
+      std::vector<double> y;
       sp2.getProfile(x, y);
       REQUIRE( x[0] == 1 );
       REQUIRE( y[0] == 3);
